added the read/write to the functions class

+++ calculations.cpp
@@ -25,8 +25,8 @@ double Calculations::variance_value(const std::vector<double>& data) { //calcula
 }
 
 int Calculations::min_value(const std::vector<double>& store_value) {
-    int min = store_value[0];
-    for (int x : store_value) {
+    double min = store_value[0];
+    for (double x : store_value) {
         if (x < min) {
             min = x;
         }
@@ -35,8 +35,8 @@ int Calculations::min_value(const std::vector<double>& store_value) {
 }
 
 int Calculations::max_value(const std::vector<double>& store_value) {
-    int max = store_value[0];
-    for (int x : store_value) {
+    double max = store_value[0];
+    for (double x : store_value) {
         if (x > max) {
             max = x;
         }
